cringe

+++ src/main/java/net/minusmc/minusbounce/injection/forge/mixins/network/MixinNetHandlerPlayClient.java
@@ -83,9 +83,9 @@ public abstract class MixinNetHandlerPlayClient {
                 float f = (float)(packetIn.getYaw() * 360) / 256.0F;
                 float f1 = (float)(packetIn.getPitch() * 360) / 256.0F;
                 EntityOtherPlayerMP entityotherplayermp = new EntityOtherPlayerMP(gameController.theWorld, getPlayerInfo(packetIn.getPlayer()).getGameProfile());
-                entityotherplayermp.prevPosX = entityotherplayermp.lastTickPosX = (double)(entityotherplayermp.serverPosX = packetIn.getX());
-                entityotherplayermp.prevPosY = entityotherplayermp.lastTickPosY = (double)(entityotherplayermp.serverPosY = packetIn.getY());
-                entityotherplayermp.prevPosZ = entityotherplayermp.lastTickPosZ = (double)(entityotherplayermp.serverPosZ = packetIn.getZ());
+                entityotherplayermp.prevPosX = entityotherplayermp.lastTickPosX = entityotherplayermp.serverPosX = packetIn.getX();
+                entityotherplayermp.prevPosY = entityotherplayermp.lastTickPosY = entityotherplayermp.serverPosY = packetIn.getY();
+                entityotherplayermp.prevPosZ = entityotherplayermp.lastTickPosZ = entityotherplayermp.serverPosZ = packetIn.getZ();
                 int i = packetIn.getCurrentItemID();
 
                 if (i == 0)
